Rough Scripts to Generate Figures for Proposal
This is a very rough scripts to generate figures for the proposal in a format which looks similar to the existing Figure produced by Lawless. Main purpose of the script is to show heterogeneity for the proposal; however it may be worth coming back to this later (e.g. for linear regressions to find growth rates).