Makefile to help Conor build Latex
Could you have a look at getting this make file to work?  It would be
great if I can build the latex too.

Idea is to type "make" at console and it runs latex (multiple times)
followed by bibtex on the relevant files and generates a .pdf.

It currently runs with some errors about bounding boxes on images and
produces .pdf with images massively enlarged.